+++ packages/railroad/src/railroad/experimental/__init__.py
@@ -6,8 +6,9 @@ compatibility and for specific use cases.
 
 Submodules:
 - environment: Legacy environment interface (EnvironmentInterface, AbstractEnvironment, etc.)
+- unknown_search: Frontier-based unknown-space exploration environment
 """
 
 from . import environment
 
-__all__ = ["environment"]
+__all__ = ["environment", "unknown_search"]
