+++ packages/railroad/tests/experimental/unknown_search/test_environment.py
@@ -8,10 +8,10 @@ import numpy as np
 
 from railroad._bindings import Fluent, State
 from railroad.navigation.constants import COLLISION_VAL, FREE_VAL
-from railroad.environment.unknown_space.environment import UnknownSpaceEnvironment
-from railroad.environment.unknown_space.types import NavigationConfig, Pose
+from railroad.experimental.unknown_search.environment import UnknownSpaceEnvironment
+from railroad.experimental.unknown_search.types import NavigationConfig, Pose
 from railroad.environment.symbolic import LocationRegistry
-from railroad.environment.unknown_space.operators import (
+from railroad.experimental.unknown_search.operators import (
     construct_move_navigable_operator,
     construct_search_at_site_operator,
 )
@@ -315,7 +315,7 @@ def test_non_interruptible_move_completes_under_aggressive_interrupt_thresholds(
 
 def test_stale_destination_is_handled_only_on_interrupt_request(monkeypatch):
     """Stale move destination is rewritten when an interrupt request is processed."""
-    import railroad.environment.unknown_space.environment as nav_env_module
+    import railroad.experimental.unknown_search.environment as nav_env_module
 
     config = NavigationConfig(
         sensor_range=9.0,
